genai-observability/app/requirements.txt
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation-openai-v2==2.3b0
